requirements.txt
sqlalchemy
python-telegram-bot
pandas
joblib
BeautifulSoup4
uvicorn
psycopg2-binary
starlette
fastapi